pyproject.toml
# ABOUTME: Package configuration for the harperreed README builder
# ABOUTME: Defines dependencies and build settings for the profile generator

[project]
name = "readme-builder"
version = "2.0.0"
description = "Self-updating GitHub profile README generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
build-readme = "readme_builder.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/readme_builder"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
